Fix front end important and minor issues (#139)
- Fix login issue with two simultaneous tabs
- Fix logout when token is invalid (401 recieved)
- Fix delete course when user information exists
- Fix close final inscription modal btn color
- Fix logout on change password view
- Remove home content
- Fix error display messages on modal
- Fix user change password error messages
- Fix alerts not displaying when already closed
- Fix getActions with undefined entity
- Add sort criteria to all lists
- Add students average on grades view
- Remove user on not found and server error views
- Fix students average translation
- Fix filters on inscriptions and add correlative
- Fix search fields and some views style and responsive behavior
- Fix numeric step on final inscription grade

+++ app/scripts/controllers/LoginCtrl.js
@@ -13,6 +13,17 @@ function(paw) {
     this.login = function(user) {
       // user would never be undefined as the form has to be valid before calling this method
       // and that means that the user has to have some value defined
+      //
+
+      // the following is for cases where the user opens a second session simulatneously on a new tab
+      if (Authentication.isLoggedIn()) {
+        // fetch the current user in case it is not already fetched
+        navDataService.fetchUser();
+        // redirect to current index
+        Paths.get().index().go();
+        return;
+      }
+
       Authentication.login(user).then(function(authToken) {
         navDataService.remove('user');
         Authentication.setToken(authToken);
